Removed Donations + Minor Changes
Removed the donation link to the salvation army and changed the text
describing what the donation was for to reflect on what happened during
the livestreams and the money raised.

Changed the title for "Stream & Donate" page to "Stream & Donations".
Updated CSS for background gradient to support resolutions > 1080p.
Reworded and fixed some grammar issues with README.md.